refactor: #118 - user 정렬 로직을 useUserSort 분리
